+++ Source/OnlineSubsystemAccelByte/Private/AsyncTasks/CloudSave/OnlineAsyncTaskAccelByteGetGameRecord.h
@@ -19,6 +19,7 @@ public:
 	FOnlineAsyncTaskAccelByteGetGameRecord(FOnlineSubsystemAccelByte* const InABInterface, int32 InLocalUserNum, const FString& InKey, bool bInAlwaysRequestToService);
 
 	virtual void Initialize() override;
+	virtual void Finalize() override;
 	virtual void TriggerDelegates() override;
 
 protected: