+++ Source/OnlineSubsystemAccelByte/Private/AsyncTasks/CloudSave/OnlineAsyncTaskAccelByteGetUserRecord.h
@@ -19,6 +19,7 @@ public:
 	FOnlineAsyncTaskAccelByteGetUserRecord(FOnlineSubsystemAccelByte* const InABInterface, const int32 InLocalUserNum, const FUniqueNetId& InLocalUserId, const FString& InKey, bool IsPublic, const FString& InRecordUserId = TEXT(""));
 
 	virtual void Initialize() override;
+	virtual void Finalize() override;
 	virtual void TriggerDelegates() override;
 
 protected: