+++ Source/OnlineSubsystemAccelByte/Private/AsyncTasks/Statistic/OnlineAsyncTaskAccelByteListUserStatItems.h
@@ -25,6 +25,7 @@ public:
 		, FString const& AdditionalKey
 		, bool bAlwaysRequestToService);
 	virtual void Initialize() override;
+	virtual void Finalize() override;
 	virtual void TriggerDelegates() override;
 
 protected: